requirements.txt
openai
python-dotenv
pyautogui
pywinauto
StopWatch.py
